Debian upstreaming - Do not install loader and setup scripts (#9139)
For upstreaming or installs into /usr or /usr/local, do not install
application loaders or setup script.  And do not install anything
under bin/unwrapped.

Rename `xbtrace` and `xbreplay` to `xrt-trace` and `xrt-replay`.
Rename library `xrt_wrapper` to `xrt_trace` and ABI version this
library.  This in turn required a code change to append the ABI
version to the name of the dlopened libxrt_trace.so library.

Do not use install RPATH when building for UPSTREAM.

Disable xclbinutil unit tests when installing to /usr or /usr/local,
apparently these tests rely on setup.sh, which is not installed for
these install prefixes.

Ensure that lib -> lib64 symlink is created if necssary when not
UPSTREAM and when install prefix is /opt/xilinx/xrt (internal builds).

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>